+++ server/db/attribution_documents.ts
@@ -6,24 +6,27 @@ import pg from './index';
 export interface AttributionDocument {
   doc_id: number;
   project_id: string;
+  project_version: string;
   created_on: Date;
+  created_by: string;
   content: string;
 }
 
 export function getAttributionDocument(
+  projectId: string,
   docId: number
 ): Promise<AttributionDocument> {
   return pg().oneOrNone(
-    'select * from attribution_documents where doc_id = $1',
-    [docId]
+    'select * from attribution_documents where project_id = $1 and doc_id = $2',
+    [projectId, docId]
   );
 }
 
 export function findDocumentsForProject(
   projectId: string
-): Promise<AttributionDocument> {
+): Promise<Pick<AttributionDocument, 'doc_id' | 'project_version' | 'created_on' | 'created_by'>[]> {
   return pg().query(
-    'select * from attribution_documents where project_id = $1 order by created_on desc',
+    'select doc_id, project_version, created_on, created_by from attribution_documents where project_id = $1 order by created_on desc',
     [projectId]
   );
 }